✨ 기능 추가 / 수정: count Integer -> Long으로 변경

+++ LearnsMate/src/main/java/intbyte4/learnsmate/report/domain/dto/ReportedMemberDTO.java
@@ -10,5 +10,5 @@ import lombok.*;
 @Builder
 public class ReportedMemberDTO {
     private Member reportedMember;
-    private Integer reportCount;
+    private Long reportCount;
 }
